Defined DataImportConsumerVerticle as bean
