+++ src/main/java/org/folio/config/ApplicationConfig.java
@@ -9,29 +9,6 @@ import org.springframework.context.annotation.Import;
 
 @Configuration
 @ComponentScan({ "org.folio" })
-@Import({ RestClientsConfiguration.class, ServicesConfiguration.class })
+@Import({ RestClientsConfiguration.class, ServicesConfiguration.class, KafkaConsumersConfiguration.class })
 public class ApplicationConfig {
-
-  @Value("${KAFKA_HOST:kafka}")
-  private String kafkaHost;
-  @Value("${KAFKA_PORT:9092}")
-  private String kafkaPort;
-  @Value("${OKAPI_URL:http://okapi:9130}")
-  private String okapiUrl;
-  @Value("${REPLICATION_FACTOR:1}")
-  private int replicationFactor;
-  @Value("${ENV:folio}")
-  private String envId;
-
-  @Bean
-  public KafkaConfig kafkaConfig() {
-    return KafkaConfig.builder()
-      .envId(envId)
-      .kafkaHost(kafkaHost)
-      .kafkaPort(kafkaPort)
-      .okapiUrl(okapiUrl)
-      .replicationFactor(replicationFactor)
-      .build();
-  }
-
 }
