+++ src/main/java/org/folio/verticles/DataImportConsumerVerticle.java
@@ -11,36 +11,45 @@ import org.folio.kafka.KafkaConsumerWrapper;
 import org.folio.kafka.KafkaTopicNameHelper;
 import org.folio.kafka.SubscriptionDefinition;
 import org.folio.processing.events.EventManager;
-import org.folio.spring.SpringContextUtil;
 import org.folio.util.pubsub.PubSubClientUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
+import org.springframework.beans.factory.config.ConfigurableBeanFactory;
+import org.springframework.context.annotation.Scope;
 import org.springframework.context.support.AbstractApplicationContext;
+import org.springframework.stereotype.Component;
 
+@Component
+@Scope(ConfigurableBeanFactory.SCOPE_PROTOTYPE)
 public class DataImportConsumerVerticle extends AbstractVerticle {
 
   private static final Logger LOGGER = LogManager.getLogger(DataImportConsumerVerticle.class);
   private static final GlobalLoadSensor GLOBAL_LOAD_SENSOR = new GlobalLoadSensor();
   public static final String EDIFACT_RECORD_CREATED_EVENT = "DI_EDIFACT_RECORD_CREATED";
 
-  private static AbstractApplicationContext springContext;
-
   @Value("${mod.invoice.kafka.DataImportConsumer.loadLimit:5}")
   private int loadLimit;
   @Value("${mod.invoice.kafka.DataImportConsumerVerticle.maxDistributionNumber:100}")
   private int maxDistributionNumber;
-  @Autowired
-  private KafkaConfig kafkaConfig;
-  @Autowired
-  private AsyncRecordHandler<String, String> dataImportKafkaHandler;
+
+  private final AbstractApplicationContext springContext;
+  private final KafkaConfig kafkaConfig;
+  private final AsyncRecordHandler<String, String> dataImportKafkaHandler;
   private KafkaConsumerWrapper<String, String> consumerWrapper;
+
+  @Autowired
+  public DataImportConsumerVerticle(KafkaConfig kafkaConfig, AsyncRecordHandler<String, String> dataImportKafkaHandler,
+                                    AbstractApplicationContext springContext) {
+    this.springContext = springContext;
+    this.kafkaConfig = kafkaConfig;
+    this.dataImportKafkaHandler = dataImportKafkaHandler;
+  }
 
   @Override
   public void start(Promise<Void> startPromise) {
-    context.put("springContext", springContext); // todo: fix it
-    SpringContextUtil.autowireDependencies(this, context);
+    context.put("springContext", springContext);
 
-    LOGGER.debug("Kafka config: {}", kafkaConfig);
+    LOGGER.info("Kafka config: {}", kafkaConfig);
     EventManager.registerKafkaEventPublisher(kafkaConfig, vertx, maxDistributionNumber);
 
     SubscriptionDefinition subscriptionDefinition = KafkaTopicNameHelper.createSubscriptionDefinition(kafkaConfig.getEnvId(),
@@ -64,7 +73,4 @@ public class DataImportConsumerVerticle extends AbstractVerticle {
     consumerWrapper.stop().onComplete(stopPromise::handle);
   }
 
-  public static void setSpringGlobalContext(AbstractApplicationContext springContext) {
-    DataImportConsumerVerticle.springContext = springContext;
-  }
 }
